+++ ChangeLog.md
@@ -8,6 +8,9 @@
 * Libs/Kotlin **(Breaking)**: Update `sendExample` to return `MessageOut` (instead of nothing)
 * Libs/Kotlin: Fix the parameter names of `Endpoint.get` - `appId` and `endpointId` were swapped
 * Libs/Kotlin: Fix a bug in `EventType.list` where `options.order` was not getting honored
+* Libs/Rust **(Breaking)**: Add optional `EventTypeDeleteOptions` parameter to `EventType::delete`
+* Libs/Rust **(Breaking)**: Add optional `PostOptions` parameter to `Endpoint::recover`,
+  `Endpoint::rotate_secret`, `Integration::rotate_key` and `MessageAttempt::resend`
 
 ## Version 1.56.0
 * Skipping versions: we had an issue with our CI that created duplicated Go